extracted edit house logic from cooperationInfo
